added function to filter kind of files in create and update controllers
